Changed implementation of AutocompleteMenu so that it no longer derives from ToolStripDropDown, but rather is just a UserControl.
ToolStripDropDown-based implementation has critical issues on Mono when working with keyboard. As soon as the autocomplete popup appears, keyboard stops working for FastColoredTextBox control, so that it is impossible to type anything. I tried different ways of programmatically returning focus to the textbox, but haven't a working solution.

On the other hand, AutocompleteListView is anyway drawn manually and doesn't actually require any ToolStripDropDown functionality to work properly, as far as I can tell. Cutting it off seems to be a pretty logical thing to do anyway.